Overhauled Json related code. Colors now written as bytes rather than floats in material json files. Removed Newtonsoft.Json for System.Text.Json

+++ source/EndianBinaryReader.cs
@@ -346,7 +346,7 @@ namespace Kai
         }
         public void Write(Bone bone)
         {
-            Write(bone.MatrixType);
+            Write((short)bone.MatrixType);
             Write(bone.InheritParentScale);
             Write((sbyte)-1);
